+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ### New Features
 
 - Add version badge to Dashboard and 'MoreMenu' (!981)
+- Show payment warning in dashboard if there is a payment issue (!982)
 - Show imprint and data protection footer in the lobby and waiting roomm (!984)
 - Show meeting title in the meeting header (!986)